+++ CHANGELOG.md
@@ -11,6 +11,7 @@ _prefixes_. This project adheres to [Semantic Versioning](https://semver.org/spe
 ### Added
 
 - Add Python 3.10 support ([#676](https://github.com/DonDebonair/slack-machine/pull/676))
+- Expose web client of Slack SDK ([#677](https://github.com/DonDebonair/slack-machine/pull/677))
 
 ## [0.31.0] - 2022-10-21